requirements.txt
# Core dependencies
pandas==2.3.2
numpy==2.3.2

# Streamlit UI Framework
streamlit==1.29.0
plotly==5.17.0

# Google Sheets API
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.179.0
httplib2==0.22.0

# PostgreSQL and Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.43

# Visualization and Charts
matplotlib==3.8.2
seaborn==0.13.0

# Additional dependencies
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
python-dotenv==1.0.0
